requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gspread==4.0.1
httplib2==0.19.1
idna==3.2
mysqlclient==2.0.3
oauth2client==4.1.3
oauthlib==3.1.1
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
simplejson==3.17.5
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.7
